Renamed test fixture and added DatabaseDriver test
